feat!: deprecate and remove legacy /api REST endpoints
